Sharing dask_client test fixture between rio and dask tests
